# cafe/resources/swagger.yml
openapi: 3.0.3
info:
  title: E-Commerce Microservices API
  description: Unified API documentation for Product, User, and Order services
  version: 1.0.0

servers:
  - url: https://api.example.com/api/v1

paths:

  ######################
  # PRODUCT APIs
  ######################

  /products:
    post:
      summary: Create product
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'

    get:
      summary: Get all products
      tags: [Products]
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

  /products/{productId}:
    get:
      summary: Get product by ID
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'

    put:
      summary: Update product
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated

    delete:
      summary: Delete product
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Product deleted

  ######################
  # USER APIs
  ######################

  /users:
    post:
      summary: Create user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created

    get:
      summary: Get all users
      tags: [Users]
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

  /users/{userId}:
    get:
      summary: Get user by ID
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    put:
      summary: Update user
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated

    delete:
      summary: Deactivate user
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deactivated

  ######################
  # ORDER APIs
  ######################

  /orders:
    post:
      summary: Create order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created

    get:
      summary: Get all orders
      tags: [Orders]
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

  /orders/{orderId}:
    get:
      summary: Get order by ID
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

  /orders/user/{userId}:
    get:
      summary: Get orders by user ID
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

components:

  ######################
  # PARAMETERS
  ######################

  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  ######################
  # SCHEMAS
  ######################

  schemas:

    ProductRequest:
      type: object
      required: [name, price, currency, quantityAvailable]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        currency:
          type: string
        quantityAvailable:
          type: integer

    ProductResponse:
      allOf:
        - $ref: '#/components/schemas/ProductRequest'
        - type: object
          properties:
            productId:
              type: string
              format: uuid
            status:
              type: string
            createdAt:
              type: string
              format: date-time

    UserRequest:
      type: object
      required: [firstName, email, password]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        phone:
          type: string

    UserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    OrderRequest:
      type: object
      required: [userId, items]
      properties:
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer

    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        orderStatus:
          type: string
        totalAmount:
          type: number
          format: double
        currency:
          type: string
        createdAt:
          type: string
          format: date-time