# order-service/src/main/java/com/microservices/order/dto/OrderRequest.java
package com.microservices.order.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;
import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class OrderRequest {

    private UUID userId;
    private List<OrderItemRequest> items;
}