# order-service/src/main/java/com/microservices/order/dto/OrderResponse.java
package com.microservices.order.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class OrderResponse {

    private UUID orderId;
    private UUID userId;
    private String orderStatus;
    private Double totalAmount;
    private String currency;
    private LocalDateTime createdAt;
}